python-service/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
pydantic
# Le fix pour Supabase/Proxy
httpx==0.27.2
# On laisse Supabase trouver la version compatible avec ce httpx
supabase
# L'IA
langgraph
langchain-openai
langchain-core
azure-ai-formrecognizer
azure-identity